TracyLite: Offline export, CMake overhaul, Perfetto support
Major update enabling TracyLite "no-send" (offline) export with both JSON (Chrome Trace) and native Perfetto formats. Modernizes the build system with comprehensive CMakePresets for all platforms, robust VSCode integration, and automatic Perfetto SDK handling (including protozero-only fast builds and offline fallback). Adds a high-performance, thread-safe ring buffer and export APIs, with runtime-selectable formats and memory tracking. Includes new examples, documentation, and a CMake helper for automatic ZoneScoped injection. All new features are opt-in and zero-overhead when disabled. Extensive source and API cleanups ensure cross-platform reliability and developer usability.

+++ profiler/src/profiler/TracyLlmEmbeddings.cpp
@@ -62,7 +62,7 @@ std::vector<TracyLlmEmbeddings::Result> TracyLlmEmbeddings::Search( const std::v
     for( size_t i=0; i<result.size(); i++ )
     {
         ret.emplace_back( Result {
-            .idx = result[i].member.key,
+            .idx = static_cast<size_t>( result[i].member.key ),
             .distance = result[i].distance
         } );
     }
